+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/integration/villagercomfort/ComfortHelperMixin.java
@@ -1,9 +1,16 @@
 package insane96mcp.iguanatweaksreborn.mixin.integration.villagercomfort;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import dev.ghen.villagercomfort.comfort.ComfortHelper;
+import dev.ghen.villagercomfort.core.config.CommonConfig;
+import net.minecraft.world.entity.ai.memory.MemoryModuleType;
+import net.minecraft.world.entity.npc.Villager;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
+
+import java.util.Optional;
 
 @Mixin(ComfortHelper.class)
 public class ComfortHelperMixin {
@@ -12,18 +19,11 @@ public class ComfortHelperMixin {
         return -1L;
     }
 
-    /*@Definition(id = "villager", local = @Local(type = Villager.class, argsOnly = true))
-    @Definition(id = "getBrain", method = "Lnet/minecraft/world/entity/npc/Villager;getBrain()Lnet/minecraft/world/entity/ai/Brain;")
-    @Definition(id = "getMemory", method = "Lnet/minecraft/world/entity/ai/Brain;getMemory(Lnet/minecraft/world/entity/ai/memory/MemoryModuleType;)Ljava/util/Optional;")
-    @Definition(id = "LAST_SLEPT", field = "Lnet/minecraft/world/entity/ai/memory/MemoryModuleType;LAST_SLEPT:Lnet/minecraft/world/entity/ai/memory/MemoryModuleType;")
-    @Definition(id = "orElse", method = "Ljava/util/Optional;orElse(Ljava/lang/Object;)Ljava/lang/Object;")
-    @Definition(id = "COMFORT_PER_DAY_WITHOUT_SLEEP", field = "Ldev/ghen/villagercomfort/core/config/CommonConfig;COMFORT_PER_DAY_WITHOUT_SLEEP:Lnet/minecraftforge/common/ForgeConfigSpec$ConfigValue;")
-    @Definition(id = "get", method = "Lnet/minecraftforge/common/ForgeConfigSpec$ConfigValue;get()Ljava/lang/Object;")
-    @Definition(id = "intValue", method = "Ljava/lang/Number;intValue()I")
-    @Definition(id = "Number", type = Number.class)
-    @Expression("(villager.getBrain().getMemory(LAST_SLEPT).orElse(0) / 24000) * ((Number) COMFORT_PER_DAY_WITHOUT_SLEEP.get()).intValue()")
-    @WrapOperation(method = "getVillagerComfort", at = @At(value = "MIXINEXTRAS:EXPRESSION", ordinal = 0), remap = false)
-    private static int iguanatweaksreborn$fixDaysWithoutSleepComfort(int original) {
-        return original;
-    }*/
+    @WrapOperation(method = "getVillagerComfort", at = @At(value = "INVOKE", target = "Ljava/util/Optional;orElse(Ljava/lang/Object;)Ljava/lang/Object;", ordinal = 1), remap = false)
+    private static Object iguanatweaksreborn$fixDaysWithoutSleepComfort(Optional<Object> instance, Object other, Operation<Object> original, Villager villager) {
+        long lastSlept = villager.getBrain().getMemory(MemoryModuleType.LAST_SLEPT).orElse(0L);
+        if (lastSlept == 0)
+            lastSlept = villager.level().getGameTime() - villager.tickCount;
+        return (villager.level().getGameTime() - lastSlept) / 24000L * CommonConfig.COMFORT_PER_DAY_WITHOUT_SLEEP.get().intValue();
+    }
 }